game: make existing size narrowing explicit and enforce C4267

+++ game/cgame/cg_draw.cpp
@@ -243,7 +243,7 @@ static void CG_DrawField (int x, int y, int width, int value) {
 	}
 
 	Com_sprintf (num, sizeof(num), "%i", value);
-	l = strlen(num);
+	l = (int)( strlen(num) );
 	if (l > width)
 		l = width;
 	x += 2 + CHAR_WIDTH*(width - l);
